requirements.txt
pynput==1.7.6
Pillow == 10.1.0
numpy == 1.26.3